Randon_Forest_Model
The Random Forest model is a powerful machine learning ensemble method that leverages multiple decision trees to make predictions. It constructs a large number of trees during training, with each tree being trained on a random subset of the data and features, which helps reduce bias and balance variance. For regression tasks, it averages the predictions of all individual trees to arrive at the final result, while for classification tasks, it uses a majority vote.